formlabel, pagination, tabs, verticalnav, footer, privacy settings, card and simplefooter updates for storybook 7
