--- packages/ds-healthcare-gov/src/components/Footer/Footer.stories.jsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import Footer from './Footer';
-
-export default {
-  title: 'Healthcare/Footer',
-  component: Footer,
-  argTypes: {
-    className: {
-      control: { type: 'text' },
-      type: { name: 'string', required: false },
-    },
-  },
-  args: {
-    className: '',
-  },
-  parameters: { theme: 'healthcare' },
-};
-
-const Template = ({ data, ...args }) => <Footer {...args} />;
-
-export const BasicFooter = Template.bind({});
